# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1-oas3", 
    "title": "Win Backend Service",
    "description": "<b>Win Project Application API</b> <br> \n\n <b>Catatan :</b> \n\n - Untuk <b>CMS</b> url menggunakan /api/v1 \n\n - Untuk <b>API</b> url menggunakan /apirest/v1"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://backendwin.addityar.xyz"
    }
  ],
	"security": [
		{
			"bearerAuth": []
		}
	],
...
  "consumes": ["application/json"],
  "produces": ["application/json"],
	"components": {
...
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"paths": {
...
		"/api/v1/settings/encryptPass": {
			"get": {
				"tags": ["CMS Settings"],
				"summary": "Untuk encrypt kata sandi",
				"parameters": [
					{
						"name": "kata",
						"in": "query",
						"style": "simple",
						"explode": false,
						"schema": {
							"type": "string",
							"example": ""
						}
					}
				],
				"responses": {
					"200": {
						"description": "Berhasil"
					},
					"404": {
						"description": "Gagal atau Unauthorized"
					}
				}
			}
		},
...
	}
}